+++ src/components/Icon/index.ts
@@ -1,4 +1,5 @@
 import Icon from './Icon';
 
-export {Props, IconSource} from './Icon';
+export {IconSource} from './types';
+export {Props} from './Icon';
 export default Icon;
